+++ client/src/constants.rs
@@ -6,11 +6,11 @@ pub const KEEP_ALIVE_DEFAULT: u64 = 5_000;
 pub const RETENTION_POLICY_DEFAULT: u64 = 0;
 
 #[cfg(debug_assertions)]
-pub(crate) const CLOUD_CA: &[u8; 469] = include_bytes!("../proxy.debug.der");
+pub(crate) const CLOUD_CA: &[u8; 390] = include_bytes!("../ca.debug.der");
 #[cfg(debug_assertions)]
 pub(crate) const SELIUM_CLOUD_REMOTE_URL: &str = "127.0.0.1:7002";
 
 #[cfg(not(debug_assertions))]
-pub(crate) const CLOUD_CA: &[u8; 470] = include_bytes!("../proxy.prod.der");
+pub(crate) const CLOUD_CA: &[u8; 391] = include_bytes!("../ca.prod.der");
 #[cfg(not(debug_assertions))]
 pub(crate) const SELIUM_CLOUD_REMOTE_URL: &str = "selium.io:7001";