+++ server/src/quic.rs
@@ -2,7 +2,6 @@
 //! `<https://github.com/quinn-rs/quinn/blob/main/quinn/examples/server.rs>`
 
 use anyhow::{bail, Context, Result};
-use openssl::x509::X509;
 use quinn::{Connection, IdleTimeout, ServerConfig};
 use rustls::server::AllowAnyAuthenticatedClient;
 use rustls::{Certificate, PrivateKey, RootCertStore};
@@ -110,7 +109,7 @@ pub fn load_root_store<T: AsRef<Path>>(ca_file: T) -> Result<RootCertStore> {
     Ok(store)
 }
 
-pub fn get_pubkey_from_connection(connection: &Connection) -> Result<String> {
+pub fn get_pubkey_from_connection(connection: &Connection) -> Result<Vec<u8>> {
     let peer_identity = connection
         .peer_identity()
         .context("Unable to read peer identity")?;
@@ -119,17 +118,9 @@ pub fn get_pubkey_from_connection(connection: &Connection) -> Result<String> {
         .downcast_ref::<Vec<Certificate>>()
         .context("Unable to read cert")?;
 
-    let pub_key = extract_public_key(certs)?;
-
-    Ok(pub_key)
-}
-
-fn extract_public_key(certs: &[Certificate]) -> Result<String> {
-    let first_cert = certs.get(0).context("Failed to get first certificate.")?;
-    let cert = X509::from_der(first_cert.as_ref())?;
-    let pub_key = cert.public_key()?;
-    let pem = pub_key.public_key_to_pem()?;
-    let result = String::from_utf8(pem)?;
-
-    Ok(result)
+    Ok(certs
+        .get(0)
+        .context("Failed to get first certificate")?
+        .0
+        .clone())
 }