+++ server/CHANGELOG.md
@@ -17,3 +17,9 @@
 - Decoupled binary into library components for testing
 - Features for Selium Cloud
 - Bump dependency versions
+
+## v0.3.1
+
+- Remove openssl dependency
+- Replace faulty Cloud certs
+- Bug fixes for Selium Cloud