+++ standard/CHANGELOG.md
@@ -9,3 +9,7 @@
 - Added connection reestablishment for failed connections
 - Features for Selium Cloud
 - Bump dependency versions
+
+## v0.2.1
+
+- Improve RequestHandlerFailure error